requirements.txt
openai>=1.0.0
elasticsearch>=8.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
pandas>=2.0.0
